delete downloaded code after parsing
